requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
albumentations
opendatasets
python-dotenv
